Use named pipes for windows processors
